Change ``admin_approval.views`` to inherit from ``default.views``
Most of the views in ``admin_approval.views`` followed the exact same
logic with the corresponding views in ``default.views`` but use
``SupervisedRegistrationProfile`` instead of ``RegistrationProfile``.
To be able to subclass those views instead of rewriting them we added
a field in each one of these storing the registration profile it uses.
This enables us to subclass the views changing only this field and
avoid repeating code.