Change order of deletion in ``RegistrationManager.delete_expired_users()`` to inherit it
Executing user.delete() first and then profile.delete() is a logical error. It was not visible
before but now it is. ``RegistrationProfile`` has a ForeignKey (actually a OneToOneField
which cascades delete) to the user, so when the user was deleted his related ``RegistrationProfile``
was deleted too and when ``SupervisedRegistrationProfile`` was deleted it yielded the following error:

self.model.MultipleObjectsReturned( DoesNotExist: RegistrationProfile matching query does not exist)